For now, we tell git to ignore the changes in platforms/ios/www
With the file .gitignore we can tell git that certain files
do not need to be tracked as part of the git. Changes in
these files are not suggested in the Github for Mac program,
for example.

In the way we use Cordova, it copies the html, css and js
over from the root folder into a folder called platforms/ios/www
files, and then creating the ios app from these copied files.

Effectively, we have the files double, and we don’t want have
to keep track of these files two times. That’s why we add
platforms/ios/www to the file .gitignore

There is an alternative way of using Cordova, in which you edit
the html/css/js seperately for each platform. If we ever
move to this way of working, we should remove that folder from
the .gitignore